requirements.txt
Package           Version
----------------- -------
annotated-types   0.7.0
anyio             4.8.0
click             8.1.8
dnspython         1.16.0
fastapi           0.115.6
greenlet          3.1.1
h11               0.14.0
idna              3.10
pip               24.0
pydantic          2.10.5
pydantic_core     2.27.2
pymongo           3.12.0
python-dotenv     1.0.1
sniffio           1.3.1
SQLAlchemy        2.0.38
starlette         0.41.3
typing_extensions 4.12.2
uvicorn           0.34.0
